pyproject.toml
[project]
name = "finance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.45",
    "uvicorn[standard]>=0.40.0",
]
